+++ Glucograph/Model/Weight+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Weight+CoreDataProperties.swift
 //  Glucograph
 //
-//  Created by Sergey Seitov on 24.07.17.
+//  Created by Sergey Seitov on 28.07.17.
 //  Copyright © 2017 V-Channel. All rights reserved.
 //
 
@@ -20,6 +20,6 @@ extension Weight {
     @NSManaged public var date: NSDate?
     @NSManaged public var recordName: String?
     @NSManaged public var synced: Bool
-    @NSManaged public var value: Int32
+    @NSManaged public var value: Double
 
 }
